distribute tests and get them to pass
