+++ java/src/jmri/jmrit/display/layoutEditor/LayoutEditorDialogs/TrackSegmentEditor.java
@@ -32,85 +32,6 @@ public class TrackSegmentEditor extends LayoutTrackEditor {
         super(layoutEditor);
     }
 
-    // ********** Members and methods found elsewhere in LayoutTrackEditors 
-    // ********** Promote to a superclass?
-
-    List<String> sensorList = new ArrayList<>();
-
-    /**
-     * Display a message describing the reason for the block selection combo box
-     * being disabled. An option is provided to hide the message. Note: The
-     * PanelMenu class is being used to satisfy the showInfoMessage requirement
-     * for a default manager type class.
-     *
-     * @since 4.11.2
-     */
-    @InvokeOnGuiThread
-    void showSensorMessage() {
-        if (sensorList.isEmpty()) {
-            return;
-        }
-        StringBuilder msg = new StringBuilder(Bundle.getMessage("BlockSensorLine1"));  // NOI18N
-        msg.append(Bundle.getMessage("BlockSensorLine2"));  // NOI18N
-        String chkDup = "";
-        sensorList.sort(null);
-        for (String sName : sensorList) {
-            if (!sName.equals(chkDup)) {
-                msg.append("<br>&nbsp;&nbsp;&nbsp; " + sName);  // NOI18N
-            }
-            chkDup = sName;
-        }
-        msg.append("<br>&nbsp;</html>");  // NOI18N
-        jmri.InstanceManager.getDefault(jmri.UserPreferencesManager.class).
-                showInfoMessage(
-                        Bundle.getMessage("BlockSensorTitle"), // NOI18N
-                        msg.toString(),
-                        "jmri.jmrit.display.PanelMenu", // NOI18N
-                        "BlockSensorMessage");  // NOI18N
-    }
-
-
-    /**
-     * Create a list of NX sensors that refer to the current layout block. This
-     * is used to disable block selection in the edit dialog. The list is built
-     * by {@link jmri.jmrit.entryexit.EntryExitPairs#layoutBlockSensors}.
-     *
-     * @since 4.11.2
-     * @param loBlk The current layout block.
-     * @return true if sensors are affected.
-     */
-    boolean hasNxSensorPairs(LayoutBlock loBlk) {
-        if (loBlk == null) {
-            return false;
-        }
-        List<String> blockSensors = InstanceManager.getDefault(jmri.jmrit.entryexit.EntryExitPairs.class)
-                .layoutBlockSensors(loBlk);
-        if (blockSensors.isEmpty()) {
-            return false;
-        }
-        sensorList.addAll(blockSensors);
-        return true;
-    }
-
-    private void addDoneCancelButtons(JPanel target, JRootPane rp, ActionListener doneCallback, ActionListener cancelCallback) {
-        // Done
-        JButton doneButton = new JButton(Bundle.getMessage("ButtonDone"));
-        target.add(doneButton);  // NOI18N
-        doneButton.addActionListener(doneCallback);
-        doneButton.setToolTipText(Bundle.getMessage("DoneHint", Bundle.getMessage("ButtonDone")));  // NOI18N
-
-        // Cancel
-        JButton cancelButton = new JButton(Bundle.getMessage("ButtonCancel")); // NOI18N
-        target.add(cancelButton);
-        cancelButton.addActionListener(cancelCallback);
-        cancelButton.setToolTipText(Bundle.getMessage("CancelHint", Bundle.getMessage("ButtonCancel")));  // NOI18N
-
-        rp.setDefaultButton(doneButton);
-        // bind ESC to close window
-        rp.getInputMap(JComponent.WHEN_ANCESTOR_OF_FOCUSED_COMPONENT).put(
-                KeyStroke.getKeyStroke(KeyEvent.VK_ESCAPE, 0), "close"); // NOI18N
-    }
-
     // ********** Members and methods from LayoutTrackEditors 
     // ********** specific to TrackSegment
 
@@ -119,7 +40,6 @@ public class TrackSegmentEditor extends LayoutTrackEditor {
 
     private JmriJFrame editTrackSegmentFrame = null;
     private final JComboBox<String> editTrackSegmentMainlineComboBox = new JComboBox<>();
-//    private JCheckBox editTrackSegmentMainlineCheckBox = new JCheckBox(Bundle.getMessage("Mainline"));
     private final JComboBox<String> editTrackSegmentDashedComboBox = new JComboBox<>();
     private final JCheckBox editTrackSegmentHiddenCheckBox = new JCheckBox(Bundle.getMessage("HideTrack"));  // NOI18N
     private final NamedBeanComboBox<Block> editTrackSegmentBlockNameComboBox = new NamedBeanComboBox<>(
