+++ java/test/jmri/jmrit/display/layoutEditor/LayoutEditorDialogs/LayoutDoubleSlipEditorTest.java
@@ -16,16 +16,6 @@ public class LayoutDoubleSlipEditorTest extends LayoutSlipEditorTest {
     public void testCtor() {
         new LayoutDoubleSlipEditor(null);
     }
-    
-    @Before
-    public void setUp() {
-        JUnitUtil.setUp();
-    }
-
-    @After
-    public void tearDown()  {
-        JUnitUtil.tearDown();
-    }
 
     // private final static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LayoutDoubleSlipEditorTest.class);
 }
