+++ java/test/jmri/jmrit/display/layoutEditor/LayoutEditorDialogs/LayoutDoubleXOverEditorTest.java
@@ -16,16 +16,6 @@ public class LayoutDoubleXOverEditorTest extends LayoutXOverEditorTest {
     public void testCtor() {
         new LayoutDoubleXOverEditor(null);
     }
-    
-    @Before
-    public void setUp() {
-        JUnitUtil.setUp();
-    }
-
-    @After
-    public void tearDown()  {
-        JUnitUtil.tearDown();
-    }
-
+ 
     // private final static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LayoutDoubleXOverEditorTest.class);
 }
