+++ java/test/jmri/jmrit/display/layoutEditor/LayoutEditorDialogs/LayoutLHXOverEditorTest.java
@@ -16,16 +16,6 @@ public class LayoutLHXOverEditorTest extends LayoutXOverEditorTest {
     public void testCtor() {
         new LayoutLHXOverEditor(null);
     }
-    
-    @Before
-    public void setUp() {
-        JUnitUtil.setUp();
-    }
-
-    @After
-    public void tearDown()  {
-        JUnitUtil.tearDown();
-    }
 
     // private final static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LayoutLHXOverEditorTest.class);
 }
