+++ java/test/jmri/jmrit/display/layoutEditor/LayoutEditorDialogs/LayoutRHXOverEditorTest.java
@@ -16,16 +16,6 @@ public class LayoutRHXOverEditorTest extends LayoutXOverEditorTest {
     public void testCtor() {
         new LayoutRHXOverEditor(null);
     }
-    
-    @Before
-    public void setUp() {
-        JUnitUtil.setUp();
-    }
-
-    @After
-    public void tearDown()  {
-        JUnitUtil.tearDown();
-    }
 
     // private final static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LayoutRHOXOverEditorTest.class);
 }
