+++ java/test/jmri/jmrit/display/layoutEditor/LayoutEditorDialogs/LayoutTurnoutEditorTest.java
@@ -17,15 +17,5 @@ public class LayoutTurnoutEditorTest extends LayoutTrackEditorTest {
         new LayoutTurnoutEditor(null);
     }
     
-    @Before
-    public void setUp() {
-        JUnitUtil.setUp();
-    }
-
-    @After
-    public void tearDown()  {
-        JUnitUtil.tearDown();
-    }
-
     // private final static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LayoutTurnoutEditorTest.class);
 }
