+++ java/test/jmri/jmrit/display/layoutEditor/LayoutEditorDialogs/LayoutTurntableEditorTest.java
@@ -17,15 +17,5 @@ public class LayoutTurntableEditorTest extends LayoutTrackEditorTest {
         new LayoutTurntableEditor(null);
     }
     
-    @Before
-    public void setUp() {
-        JUnitUtil.setUp();
-    }
-
-    @After
-    public void tearDown()  {
-        JUnitUtil.tearDown();
-    }
-
     // private final static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LayoutTurntableEditorTest.class);
 }
