+++ java/test/jmri/jmrit/display/layoutEditor/LayoutEditorDialogs/LevelXingEditorTest.java
@@ -17,15 +17,5 @@ public class LevelXingEditorTest extends LayoutTrackEditorTest {
         new LevelXingEditor(null);
     }
     
-    @Before
-    public void setUp() {
-        JUnitUtil.setUp();
-    }
-
-    @After
-    public void tearDown()  {
-        JUnitUtil.tearDown();
-    }
-
     // private final static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LayoutTrackEditorTest.class);
 }
