+++ java/test/jmri/jmrit/display/layoutEditor/LayoutEditorDialogs/TrackSegmentEditorTest.java
@@ -17,16 +17,5 @@ public class TrackSegmentEditorTest extends LayoutTrackEditorTest {
         new TrackSegmentEditor(null);
     }
     
-    @Before
-    public void setUp() {
-        JUnitUtil.setUp();
-    }
-
-    @After
-    public void tearDown()  {
-        JUnitUtil.clearShutDownManager();
-        JUnitUtil.tearDown();
-    }
-
     // private final static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(TrackSegmentEditorTest.class);
 }
